+++ spec/style_helper_spec.rb
@@ -33,29 +33,4 @@ describe '#style_helper' do
       expect(webdriver.get_style_parameter('//div', 'foo')).to be_nil
     end
   end
-
-  describe '#set_style_show_by_xpath' do
-    it 'set_style_show_by_xpath by default show object' do
-      webdriver.set_style_show_by_xpath('//span')
-      expect(webdriver.get_style_parameter('//span', 'display')).to eq('block')
-    end
-
-    describe 'set_style_show_by_xpath with move_to_center' do
-      before do
-        webdriver.set_style_show_by_xpath('//span', true)
-      end
-
-      it 'object is shown' do
-        expect(webdriver.get_style_parameter('//span', 'display')).to eq('block')
-      end
-
-      it 'object is moved by x' do
-        expect(webdriver.get_style_parameter('//span', 'left')).to eq('410px')
-      end
-
-      it 'object is moved by y' do
-        expect(webdriver.get_style_parameter('//span', 'top')).to eq('260px')
-      end
-    end
-  end
 end