+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Changes
 
 * Extract `WebDriver#set_style_attribute` test to separate spec
+* Extract `WebDriver#set_style_show_by_xpath` test to separate spec
 
 ## 0.3.4 (2020-07-06)